Track gen_stream processes directly for accurate test results
The current test suite relies on erlang:system_info(process_count) to
determine if the requested number of worker processes is used
internally for serving up the stream. This is an unreliable method
because unrelated processes may come and go during the test, resulting
in spurious and unpredictable test failures.

This fix involves exposing the number of processes, and the actual
process list, using a new interface 'gen_stream:proc_info/1'.  This is
a very dangerous and undesirable interface for the general use of
gen_stream but it is invaluable in verifying the internal workings of
a gen_stream server during the test suite and provides a direct and
consistent way of verifying not only the number of processes used, but
whether the originally spawned processes are still active and
guaranteeing that they are not replaced by other worker processes.