Dialyzer: add --fullpath option for printing full paths in warnings
Currently, Dialyzer warning messages only contain the base name of the
file, e.g.:

mysql_conn.erl:648: The created fun has no local return

However, when running Dialyzer on a deep directory tree it is
sometimes useful to have the full paths in the warning output.  In
particular, when viewing the output in Emacs, hitting RET on an entry
to jump to the error will often yield a prompt for which directory the
file is in; with full paths we would avoid that prompt.

This change adds a --fullpath option to Dialyzer, which makes the
warning messages contain the full path of the corresponding file.